backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4
gunicorn>=20.0.4
python-dotenv>=0.15.0